RBE10 calculation adjustments
I re-implemented the LQ fit with a function already provided in the GSL library. This must be installed and I also added three lines in the CMakeLists.txt to include and link the libraries (redo the cmake $path_to_monas_standalone). However, this was not the problem for the calculation of RBE10, the linear fit was overwriting the alpha on the LQ one. Now it works fine.